+++ src/components/ui/badge.tsx
@@ -41,4 +41,4 @@ function Badge({ className, variant, size, ...props }: BadgeProps) {
   return <div className={cn(badgeVariants({ variant, size }), className)} {...props} />;
 }
 
-export { Badge, badgeVariants };
+export { Badge };
